+++ src/OrchardCore.Modules/OrchardCore.Sms/Views/_ViewImports.cshtml
@@ -1,8 +1,3 @@
-@*
-    For more information on enabling MVC for empty projects, visit http://go.microsoft.com/fwlink/?LinkID=397860
-
-*@
-
 @inherits OrchardCore.DisplayManagement.Razor.RazorPage<TModel>
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
