+++ src/OrchardCore.Modules/OrchardCore.Users/ViewModels/EnableSmsAuthenticatorViewModel.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel.DataAnnotations;
-using Microsoft.AspNetCore.Mvc.ModelBinding;
 
 namespace OrchardCore.Users.ViewModels;
 
@@ -9,12 +8,4 @@ public class EnableSmsAuthenticatorViewModel
     public string Code { get; set; }
 
     public string PhoneNumber { get; set; }
-}
-
-public class RequestCodeSmsAuthenticatorViewModel
-{
-    public string PhoneNumber { get; set; }
-
-    [BindNever]
-    public bool AllowChangePhoneNumber { get; set; }
 }
