+++ src/OrchardCore.Modules/OrchardCore.Users/ViewModels/RequestCodeSmsAuthenticatorViewModel.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.AspNetCore.Mvc.ModelBinding;
+
+namespace OrchardCore.Users.ViewModels;
+
+public class RequestCodeSmsAuthenticatorViewModel
+{
+    public string PhoneNumber { get; set; }
+
+    [BindNever]
+    public bool AllowChangePhoneNumber { get; set; }
+}
