+++ src/OrchardCore/OrchardCore.Users.Core/Models/EmailAuthenticatorLoginSettings.cs
@@ -1,5 +1,3 @@
-using System.ComponentModel;
-
 namespace OrchardCore.Users.Models;
 
 public class EmailAuthenticatorLoginSettings
@@ -9,14 +7,6 @@ public class EmailAuthenticatorLoginSettings
     public const string DefaultBody = "We received your request for a one-time code to authentication your account. Your code is {{ Code }}";
 
     public string Subject { get; set; }
-
-    public string Body { get; set; }
-}
-
-public class SmsAuthenticatorLoginSettings
-{
-    [DefaultValue(true)]
-    public bool AllowChangePhoneNumber { get; set; } = true;
 
     public string Body { get; set; }
 }
